refactor(channels): separate outbound echo state
